+++ infrastructure/product-persistence/src/main/java/com/commerce/product/infrastructure/persistence/repository/ProductJpaRepositoryCustom.java
@@ -5,14 +5,15 @@ import com.commerce.product.infrastructure.persistence.entity.ProductJpaEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public interface ProductJpaRepositoryCustom {
     Page<ProductJpaEntity> searchProducts(
         String categoryId,
         String keyword,
-        Integer minPrice,
-        Integer maxPrice,
+        BigDecimal minPrice,
+        BigDecimal maxPrice,
         Set<ProductStatus> statuses,
         Pageable pageable
     );
